+++ addons/explosives/config.cpp
@@ -40,13 +40,4 @@ class CfgMineTriggers {
     };
 };
 
-class ACE_Settings {
-    class GVAR(RequireSpecialist) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(PunishNonSpecialists) {
-        value = 1;
-        typeName = "BOOL";
-    };
-};
+#include "ACE_Settings.hpp"
