+++ addons/maptools/ACE_Settings.hpp
@@ -1,0 +1,6 @@
+class ACE_Settings {
+    class GVAR(EveryoneCanDrawOnBriefing) {
+        value = 1;
+        typeName = "BOOL";
+    };
+};
