+++ addons/optionsmenu/ACE_Settings.hpp
@@ -1,0 +1,9 @@
+class ACE_Settings {
+    class GVAR(optionMenuDisplaySize) {
+        value = 0;
+        typeName = "SCALAR";
+        isClientSettable = 1;
+        displayName = "$STR_ACE_optionsMenu_uiScaing";
+        values[] = {"$str_medium", "$str_large", "$str_very_large"};
+    };
+};
