+++ addons/overheating/config.cpp
@@ -20,13 +20,5 @@ class CfgPatches {
 
 #include "CfgWeapons.hpp"
 
-class ACE_Settings {
-    class GVAR(DisplayTextOnJam) {
-        typeName = "BOOL";
-        isClientSettable = 1;
-        value = 1;
-        displayName = "$STR_ACE_overheating_SettingDisplayTextName";
-        description = "$STR_ACE_overheating_SettingDisplayTextDesc";
-    };
-};
+#include "ACE_Settings.hpp"
