+++ addons/reload/config.cpp
@@ -22,13 +22,4 @@ class CfgPatches {
 
 #include "RscInGameUI.hpp"
 
-class ACE_Settings {
-    class GVAR(DisplayText) {
-        typeName = "BOOL";
-        isClientSettable = 1;
-        value = 1;
-        displayName = "$STR_ACE_reload_SettingDisplayTextName";
-        description = "$STR_ACE_reload_SettingDisplayTextDesc";
-    };
-};
-
+#include "ACE_Settings.hpp"
