+++ addons/respawn/ACE_Settings.hpp
@@ -1,0 +1,14 @@
+class ACE_Settings {
+    class GVAR(SavePreDeathGear) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(RemoveDeadBodiesDisconnected) {
+        value = 1;
+        typeName = "BOOL";
+    };
+    class GVAR(BodyRemoveTimer) {
+        value = 0;
+        typeName = "SCALAR";
+    };
+};
