+++ addons/respawn/config.cpp
@@ -16,18 +16,4 @@ class CfgPatches {
 #include "CfgAddons.hpp"
 #include "CfgVehicleClasses.hpp"
 #include "CfgVehicles.hpp"
-
-class ACE_Settings {
-    class GVAR(SavePreDeathGear) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(RemoveDeadBodiesDisconnected) {
-        value = 1;
-        typeName = "BOOL";
-    };
-    class GVAR(BodyRemoveTimer) {
-        value = 0;
-        typeName = "SCALAR";
-    };
-};
+#include "ACE_Settings.hpp"
