+++ addons/switchunits/ACE_Settings.hpp
@@ -1,0 +1,30 @@
+class ACE_Settings {
+    class GVAR(EnableSwitchUnits) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(SwitchToWest) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(SwitchToEast) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(SwitchToIndependent) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(SwitchToCivilian) {
+        value = 0;
+        typeName = "BOOL";
+    };
+    class GVAR(EnableSafeZone) {
+        value = 1;
+        typeName = "BOOL";
+    };
+    class GVAR(SafeZoneRadius) {
+        value = 100;
+        typeName = "SCALAR";
+    };
+};
