+++ addons/switchunits/config.cpp
@@ -14,34 +14,4 @@ class CfgPatches {
 
 #include "CfgEventHandlers.hpp"
 #include "CfgVehicles.hpp"
-
-class ACE_Settings {
-    class GVAR(EnableSwitchUnits) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(SwitchToWest) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(SwitchToEast) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(SwitchToIndependent) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(SwitchToCivilian) {
-        value = 0;
-        typeName = "BOOL";
-    };
-    class GVAR(EnableSafeZone) {
-        value = 1;
-        typeName = "BOOL";
-    };
-    class GVAR(SafeZoneRadius) {
-        value = 100;
-        typeName = "SCALAR";
-    };
-};
+#include "ACE_Settings.hpp"
